dom/media/test/test_mediastreamtrack_as_eventarget.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test MediaStreamTrack as EventTarget</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
  <script>
    // The test needs to access a MediaStreamTrack in some way.
    var mst = (new MediaStreamAudioDestinationNode(new AudioContext())).stream.getTracks()[0];
    var listenerForFooWasCalled = false;
    mst.addEventListener("foo", () => { listenerForFooWasCalled = true; });
    var listenerForMuteWasCalled = false;
    mst.addEventListener("mute", () => { listenerForMuteWasCalled = true; });
    var handlerForMuteWasCalled = false;
    mst.onmute = () => { handlerForMuteWasCalled = true; };

    mst.dispatchEvent(new Event("foo"));
    mst.dispatchEvent(new Event("mute"));

    ok(listenerForFooWasCalled,
       "Should have called the event listener for 'foo'");
    ok(listenerForMuteWasCalled,
       "Should have called the event listener for 'mute'");
    ok(handlerForMuteWasCalled,
       "Should have called the event handler for 'mute'");
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
</body>
</html>
